% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{create_tree_chain}
\alias{create_tree_chain}
\title{create_tree_chain
from cladistic information on several species, builds the tree chain}
\usage{
create_tree_chain(data)
}
\arguments{
\item{data}{a matrix containing cladistic inforamtion on the species}
}
\value{
a tree chain
}
\description{
create_tree_chain
from cladistic information on several species, builds the tree chain
}
