% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{dat_prep_phylo}
\alias{dat_prep_phylo}
\title{dat_prep_phylo
intermediate function before computing hill numbers
sort and name the leaves and nodes of the tree
assign a legnth if there is no branch length in the tree}
\usage{
dat_prep_phylo(comm, tree)
}
\arguments{
\item{comm}{}

\item{tree}{}
}
\value{

}
\description{
dat_prep_phylo
intermediate function before computing hill numbers
sort and name the leaves and nodes of the tree
assign a legnth if there is no branch length in the tree
}
