% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{FD_MLE}
\alias{FD_MLE}
\title{FD_MLE
R code for obtaining functional diversity (FD) based on Chao et al. (2018)
Please cite Chao et al. (2018) An attribute-diversity approach to functional diversity, functional beta diversity, and related (dis)similarity measures.
Functional Diversity of a single site for specified values of tau and q
FD_MLE (data, dij, tau, q) is a function of obtaining FD index of order q.}
\usage{
FD_MLE(data, dij, tau, q)
}
\arguments{
\item{data}{a vector of species sample frequencies.}

\item{dij}{a matrix of species-pairwise distances.}

\item{tau}{a numeric for a specified level of threshold distinctiveness.}

\item{q}{a numeric for a specified diversity order q.}
}
\value{
a numeric value of FD.
}
\description{
FD_MLE
R code for obtaining functional diversity (FD) based on Chao et al. (2018)
Please cite Chao et al. (2018) An attribute-diversity approach to functional diversity, functional beta diversity, and related (dis)similarity measures.
Functional Diversity of a single site for specified values of tau and q
FD_MLE (data, dij, tau, q) is a function of obtaining FD index of order q.
}
