% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{find_bestq}
\alias{find_bestq}
\title{find_bestq
This function computes the hill numbers for TD/PD/FD with q_nb different q between q_min and q_max
it selects the q for wich the r² of lm(esth_score~hill_number) is maximized and finally compute the
hill numbers for this q}
\usage{
find_bestq(
  div,
  data,
  tree = NULL,
  dij = NULL,
  tau = NULL,
  esth,
  q_min,
  q_max,
  q_nb,
  ncores,
  log_trans = FALSE
)
}
\arguments{
\item{div}{the diversity worked with : c(TD,PD,FD)}

\item{data}{matrix of relative abundance. Quadrat_code must be rownames and Species_name must be colnames}

\item{tree}{phylotree generated by ape::read_tree. Default is NULL. Only needed if div = PD}

\item{dij}{distance matrix generated by ade4::dist.kab (Pavoine et al. 2009). Default is NULL.}

\item{tau}{threshold for dij (Chao et al. 2019). Default is NULL, recommended value is mean(dij).}

\item{esth}{a two columns matrix of the aesthetic scores. rownames(esth) must be the same than rownames(data).}

\item{q_min}{min value for q}

\item{q_max}{max value for q}

\item{q_nb}{number of values to test for q in range q_min;q_max}

\item{ncores}{number of cores to use}

\item{log_trans}{logical. Default is FALSE. if TRUE the tested relation is lm(log(esth)~log(hill_index))}
}
\value{
hill_number_q -> vector of the hill number with q maximizing the R² for all assemblages
bestq         -> the q maximizing the R² of lm(esth_score~hill_number)
bestr2        -> max value of R²
all_r2_q      -> all q value tested and the associated R²
}
\description{
find_bestq
This function computes the hill numbers for TD/PD/FD with q_nb different q between q_min and q_max
it selects the q for wich the r² of lm(esth_score~hill_number) is maximized and finally compute the
hill numbers for this q
}
