% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hh}
\alias{hh}
\title{hh
shorter for the here::here function
see here::here for more info}
\usage{
hh(path, ...)
}
\arguments{
\item{path}{path to file}

\item{...}{}
}
\value{
path to file
}
\description{
hh
shorter for the here::here function
see here::here for more info
}
