% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{hill_taxa}
\alias{hill_taxa}
\title{hill_taxa
computes the taxonomic hill number of each row of a gic-ven matrix
with a given value of q}
\usage{
hill_taxa(comm, q = 0, MARGIN = 1, base = exp(1))
}
\arguments{
\item{comm}{a matrix with assemblages in row, species relative abundance in column}

\item{q}{}

\item{MARGIN}{}

\item{base}{}
}
\value{
a vector of TD
}
\description{
hill_taxa
computes the taxonomic hill number of each row of a gic-ven matrix
with a given value of q
}
