% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{look_cor}
\alias{look_cor}
\title{look_cor
This function orders the species according to their individual effect on the aesthetic score}
\usage{
look_cor(all_id, dat_cor, respvar, ord)
}
\arguments{
\item{all_id}{}

\item{dat_cor}{}

\item{respvar}{}

\item{ord}{}
}
\description{
look_cor
This function orders the species according to their individual effect on the aesthetic score
}
