% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reduce_mod}
\alias{reduce_mod}
\title{reduce_mod
reduces the model by removing the non-significant species}
\usage{
reduce_mod(sp_list, respvar, dat_cor, thr)
}
\arguments{
\item{sp_list}{}

\item{respvar}{}

\item{dat_cor}{}

\item{thr}{}
}
\value{

}
\description{
reduce_mod
reduces the model by removing the non-significant species
}
