% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hill.R
\name{relab_hill_phylo}
\alias{relab_hill_phylo}
\title{relab_hill_phylo
Computes the hill number of assemblages}
\usage{
relab_hill_phylo(
  comm,
  tree,
  q = 0,
  base = exp(1),
  rel_then_pool = TRUE,
  show.warning = TRUE
)
}
\arguments{
\item{tree}{phylo tree of all the species considered}

\item{q}{value of q with which Hill numbers must be computed}

\item{base}{}

\item{rel_then_pool}{}

\item{show.warning}{}

\item{comma}{matrix containing the relative abundance of the species in the assemblages}
}
\value{
a vector of PD
}
\description{
relab_hill_phylo
Computes the hill number of assemblages
}
