% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standefsize}
\alias{standefsize}
\title{standefsize
This function generates nb null assemblages of each possible number of species in the data table
computes the mean and sd of the div index on these assemblages
computes the deviation between the observed div and the SES_div}
\usage{
standefsize(
  div,
  data,
  tree = NULL,
  dij = NULL,
  tau = NULL,
  q,
  hilldiv,
  log_trans = FALSE,
  nb = 10,
  mcores = 20
)
}
\arguments{
\item{div}{the diversity worked with : c(PD,FD)
if div = PD, hillfun = relab_hill_phylo
if div = FD, hillfun = FD_MLE}

\item{data}{matrix of relative abundance. Quadrat_code must be rownames and Species_name must be colnames}

\item{tree}{phylotree generated by ape::read_tree. Default is NULL. Only needed if div = PD}

\item{dij}{distance matrix generated by ade4::dist.kab (Pavoine et al. 2009). Default is NULL. Only needed if div = FD}

\item{tau}{threshold for dij (Chao et al. 2019). Default is NULL, recommended value is mean(dij). Only needed if div = FD}

\item{q}{the value of q to use, same used to compute the "observed" PD or FD}

\item{hilldiv}{vector of hillnumber (qPD or qFD) for all assemblages}

\item{log_trans}{logical. defaukt is FALSE. if TRUE, the SES if computed on the log of the index}

\item{nb}{number of null models to generate (9999 is default)}

\item{mcores}{number of cores to use}
}
\value{
a matrix of three columns : observed hillnumber, nb of species, SES hill number
}
\description{
standefsize
This function generates nb null assemblages of each possible number of species in the data table
computes the mean and sd of the div index on these assemblages
computes the deviation between the observed div and the SES_div
}
